requirements.txt
scipy==1.0.1
numpy==1.14.2
pypng==0.0.18
requests==2.18.4